refactor: 코멘트 수정사항 반영
