+++ build.gradle
@@ -97,7 +97,7 @@ project(":layer-common") {
 
 	dependencies {
 		implementation 'org.springframework.boot:spring-boot-starter-web'
-		implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+//		implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	}
 
 }
