+++ layer-api/src/main/java/org/layer/domain/jwt/exception/TokenException.java
@@ -1,4 +1,7 @@
-package org.layer.domain.auth.exception;
+package org.layer.domain.jwt.exception;
+
+
+
 
 // 임시로 만들어놓은 Exception입니다
 public class TokenException extends RuntimeException {
